Passing range-raptor test
