+++ test/Project.toml
@@ -11,4 +11,6 @@ SnapshotTests = "4acb3128-ab29-4f79-81e6-a8e25fdcc44d"
 StatsBase = "2913bbd2-ae8a-5f71-8c99-4fb6c76f3a91"
 StructEquality = "6ec83bb0-ed9f-11e9-3b4c-2b04cb4e219c"
 Test = "8dfed614-e22c-5e08-85e1-65c5234f0b40"
+TestItemRunner = "f8b46487-2199-4994-9208-9a1283c18c0a"
+TestItems = "1c621080-faea-4a02-84b6-bbd5e436b8fe"
 ZipFile = "a5390f91-8eb1-5f08-bee0-b1d1ffed6cea"
